localization and move code
