+++ src/Cli/dotnet/commands/dotnet-test/LocalizableStrings.resx
@@ -460,4 +460,27 @@ For MSTest before 2.2.4, the timeout is used for all testcases.</value>
   <data name="CmdTestOutputDescription" xml:space="preserve">
     <value>Verbosity of test output.</value>
   </data>
+  <data name="HelpExtensionOptions" xml:space="preserve">
+    <value>Extension Options:</value>
+  </data>
+  <data name="HelpModuleIsMissingTheOptionBelow" xml:space="preserve">
+    <value>{0} is missing the option below</value>
+    <comment>0 is name of the module, this string is followed by lined that list the option</comment>
+  </data>
+  <data name="HelpModuleIsMissingTheOptionsBelow" xml:space="preserve">
+    <value>{0} is missing the options below</value>
+    <comment>0 is name of the module, this string is followed by lined that list the option</comment>
+  </data>
+  <data name="HelpOptions" xml:space="preserve">
+    <value>Options:</value>
+  </data>
+  <data name="HelpUnavailableExtensionOptions" xml:space="preserve">
+    <value>Unavailable extension options:</value>
+  </data>
+  <data name="HelpUnavailableOptions" xml:space="preserve">
+    <value>Unavailable options:</value>
+  </data>
+  <data name="HelpWaitingForOptionsAndExtensions" xml:space="preserve">
+    <value>Waiting for options and extensions...</value>
+  </data>
 </root>
