+++ src/Cli/dotnet/commands/dotnet-test/Terminal/TerminalTestReporter.cs
@@ -4,6 +4,7 @@
 using System.Collections.Concurrent;
 using System.Globalization;
 using System.Text.RegularExpressions;
+using Microsoft.DotNet.Cli;
 using Microsoft.Testing.Platform.Helpers;
 using LocalizableStrings = Microsoft.DotNet.Tools.Test.LocalizableStrings;
 
@@ -1035,5 +1036,60 @@ internal sealed partial class TerminalTestReporter : IDisposable
         }
 
         _terminalWithProgress.UpdateWorker(asm.SlotIndex);
+    }
+
+    internal void WriteHelpOptions(ConcurrentDictionary<string, CommandLineOption> commandLineOptionNameToModuleNames, Dictionary<bool, List<DotNet.Cli.CommandLineOption>> allOptions, Dictionary<bool, List<(string, string[])>> moduleToMissingOptions)
+    {
+        WriteOptionsToConsole(commandLineOptionNameToModuleNames, allOptions);
+        WriteModulesToMissingOptionsToConsole(moduleToMissingOptions);
+    }
+
+    private void WriteOptionsToConsole(ConcurrentDictionary<string, CommandLineOption> commandLineOptionNameToModuleNames, Dictionary<bool, List<CommandLineOption>> options)
+    {
+        int maxOptionNameLength = commandLineOptionNameToModuleNames.Keys.ToArray().Max(option => option.Length);
+
+        foreach (KeyValuePair<bool, List<CommandLineOption>> optionGroup in options)
+        {
+            WriteMessage(string.Empty);
+            WriteMessage(optionGroup.Key ? LocalizableStrings.HelpOptions : LocalizableStrings.HelpExtensionOptions);
+
+            foreach (CommandLineOption option in optionGroup.Value)
+            {
+                WriteMessage($"{new string(' ', 2)}--{option.Name}{new string(' ', maxOptionNameLength - option.Name.Length)} {option.Description}");
+            }
+        }
+    }
+
+    private void WriteModulesToMissingOptionsToConsole(Dictionary<bool, List<(string, string[])>> modulesWithMissingOptions)
+    {
+        var yellow = new SystemConsoleColor { ConsoleColor = ConsoleColor.Yellow };
+        foreach (KeyValuePair<bool, List<(string, string[])>> groupedModules in modulesWithMissingOptions)
+        {
+            WriteMessage(string.Empty);
+            WriteMessage(groupedModules.Key ? LocalizableStrings.HelpUnavailableOptions : LocalizableStrings.HelpUnavailableExtensionOptions, yellow);
+
+            foreach ((string module, string[] missingOptions) in groupedModules.Value)
+            {
+                if (module.Length == 0)
+                {
+                    continue;
+                }
+
+                StringBuilder line = new();
+                for (int i = 0; i < missingOptions.Length; i++)
+                {
+                    if (i == missingOptions.Length - 1)
+                        line.Append($"--{missingOptions[i]}");
+                    else
+                        line.Append($"--{missingOptions[i]}\n");
+                }
+
+                string format = missingOptions.Length == 1
+                    ? LocalizableStrings.HelpModuleIsMissingTheOptionBelow
+                    : LocalizableStrings.HelpModuleIsMissingTheOptionsBelow;
+                var missing = string.Format(format, module);
+                WriteMessage($"{missing}\n{line}\n");
+            }
+        }
     }
 }
